Address review feedback.
Specifically:
  1. Move the net package to util/testhelpers.
  2. Remove the concurrency test.
  3. Rename the SignerServer listner member.
